Lesson 13 (#14)
* Lesson 13

* Lesson 13 - update after checkstyle

* Lesson 13 - update after checkstyle

* Lesson 13 - update with proper .yml files

* Lesson 13 - update with proper .yml files _ v2

* Lesson 13 - update with proper .yml files _ v2

* Lesson 13 - update with proper .yml files _ v3

---------

+++ app/src/main/java/pl/futurecollars/invoicing/db/file/IdService.java
@@ -9,7 +9,7 @@ public class IdService {
 
   private final FilesService filesService;
   private final Path idFilePath;
-  private int nextId = 1;
+  private long nextId = 1;
 
   public IdService(Path idFilePath, FilesService filesService) {
     this.idFilePath = idFilePath;
@@ -27,7 +27,7 @@ public class IdService {
     }
   }
 
-  public int getNextIdAndIncrement() {
+  public long getNextIdAndIncrement() {
     try {
       filesService.writeSingleLineToFile(idFilePath, String.valueOf(nextId + 1));
       return nextId++;
